requirements.txt
django
stripe
python-decouple